feat: implement FinalMatch and IntermediateMatch components for tournament display
