+++ src/pages/GamePage/TournamentPage/components/MatchNode/index.tsx
@@ -1,7 +1,9 @@
 import type { Match, TournamentRoundType, VerticalPosition, HorizontalVariant } from '@/api';
 
-import { MatchLines } from '../MatchLines';
-import { PlayerCard } from '../PlayerCard';
+import { FinalMatch } from './FinalMatch';
+import { IntermediateMatch } from './IntermediateMatch';
+import { LeafMatch } from './LeafMatch';
+import { SemiFinalDecoration } from './SemiFinalDecoration';
 import * as styles from './styles.css';
 import {
   TOURNAMENT_ROUNDS,
@@ -26,10 +28,6 @@ type MatchTreePosition = {
   centerSide: HorizontalVariant;
   shouldRenderLeftLine: boolean;
   shouldRenderRightLine: boolean;
-};
-
-const isPlayerReady = (playerId: string | undefined, readyPlayerIds: number[]): boolean => {
-  return !!playerId && readyPlayerIds.includes(parseInt(playerId, 10));
 };
 
 const isPlayerInMatch = (match: Match | undefined, playerId: string): boolean => {
@@ -129,43 +127,40 @@ export const MatchNode = ({
     match.winnerId === leftSemiFinal?.winnerId ? WINNER_POSITIONS.TOP : WINNER_POSITIONS.BOTTOM;
 
   if (!hasChildren) {
-    const isFinalRoundLeaf = currentRound === TOURNAMENT_ROUNDS.FINAL;
-    const showStats = currentRound === TOURNAMENT_ROUNDS.SEMI_FINAL;
-    const player1Eliminated = isFinalRoundLeaf && match.player1?.id !== match.winnerId;
-    const player2Eliminated = isFinalRoundLeaf && match.player2?.id !== match.winnerId;
-
     return (
-      <div className={isTwoPlayer ? styles.leafWrapperHorizontal : styles.leafWrapper}>
-        <PlayerCard
-          player={match.player1}
-          isWinner={match.player1?.id === match.winnerId}
-          isLoser={player1Eliminated}
-          isReady={isPlayerReady(match.player1?.id, readyIds)}
-          showStats={showStats}
-          round={currentRound}
-          side={side}
-        />
-        <MatchLines
-          winner={
-            match.winnerId === match.player1?.id ? WINNER_POSITIONS.TOP : WINNER_POSITIONS.BOTTOM
-          }
-          variant={side ?? PLAYER_POSITIONS.LEFT}
-          highlight={isRound2}
-          round={currentRound}
-          isTwoPlayer={isTwoPlayer}
-        />
-        <PlayerCard
-          player={match.player2}
-          isWinner={match.player2?.id === match.winnerId}
-          isLoser={player2Eliminated}
-          isReady={isPlayerReady(match.player2?.id, readyIds)}
-          showStats={showStats}
-          round={currentRound}
-          side={side}
-        />
-      </div>
+      <LeafMatch
+        match={match}
+        readyIds={readyIds}
+        currentRound={currentRound}
+        side={side}
+        isTwoPlayer={isTwoPlayer}
+      />
     );
   }
+
+  const renderCenterContent = () => {
+    if (shouldRenderSemiFinalDecoration) {
+      return <SemiFinalDecoration />;
+    }
+
+    if (shouldRenderFinalMatch && match.children) {
+      return <FinalMatch match={match} readyIds={readyIds} semiFinals={match.children} />;
+    }
+
+    return (
+      <IntermediateMatch
+        match={match}
+        readyIds={readyIds}
+        currentRound={currentRound}
+        side={side}
+        isRoot={isRoot}
+        isRound4={isRound4}
+        winnerFromLeft={winnerFromLeft}
+        matchTreePosition={matchTreePosition}
+        showLine={showLine}
+      />
+    );
+  };
 
   return (
     <div className={styles.treeWrapper}>
@@ -181,121 +176,7 @@ export const MatchNode = ({
               tournamentSize={tournamentSize}
             />
           )}
-          <div className={styles.centerColumn}>
-            {shouldRenderSemiFinalDecoration ? (
-              <div className={styles.centerDecoration}>
-                <img
-                  src="/assets/images/tournament-arrow.png"
-                  alt="Arrow Left"
-                  className={styles.arrowLeft}
-                />
-                <img
-                  src="/assets/images/tournament-placeholder-ball.png"
-                  alt="Mystery Ball"
-                  className={styles.questionImage}
-                />
-                <img
-                  src="/assets/images/tournament-arrow.png"
-                  alt="Arrow Right"
-                  className={styles.arrowRight}
-                />
-              </div>
-            ) : shouldRenderFinalMatch ? (
-              (() => {
-                const [semiFinal1, semiFinal2] = match.children ?? [];
-                const eliminatedPlayerIds = [semiFinal1, semiFinal2].flatMap((semiFinal) => {
-                  if (!semiFinal?.player1 || !semiFinal?.player2 || !semiFinal.winnerId) return [];
-                  return semiFinal.player1.id === semiFinal.winnerId
-                    ? [semiFinal.player2.id]
-                    : [semiFinal.player1.id];
-                });
-
-                return (
-                  <div className={styles.matchBoxNoBorder}>
-                    <PlayerCard
-                      player={match.player1}
-                      isWinner={match.player1?.id === match.winnerId}
-                      isLoser={!!match.player1 && eliminatedPlayerIds.includes(match.player1.id)}
-                      isReady={isPlayerReady(match.player1?.id, readyIds)}
-                      showStats
-                      round={TOURNAMENT_ROUNDS.FINAL}
-                      side={PLAYER_POSITIONS.LEFT}
-                    />
-                    <span className={styles.vs}>vs</span>
-                    <PlayerCard
-                      player={match.player2}
-                      isWinner={match.player2?.id === match.winnerId}
-                      isLoser={!!match.player2 && eliminatedPlayerIds.includes(match.player2.id)}
-                      isReady={isPlayerReady(match.player2?.id, readyIds)}
-                      showStats
-                      round={TOURNAMENT_ROUNDS.FINAL}
-                      side={PLAYER_POSITIONS.RIGHT}
-                    />
-                  </div>
-                );
-              })()
-            ) : (
-              (() => {
-                const isFinalRound = currentRound === TOURNAMENT_ROUNDS.FINAL;
-                const showStats = currentRound === TOURNAMENT_ROUNDS.SEMI_FINAL;
-                const player1Eliminated = match.player1?.id !== match.winnerId;
-                const player2Eliminated = match.player2?.id !== match.winnerId;
-
-                return (
-                  <>
-                    {showLine && (
-                      <MatchLines
-                        winner={winnerFromLeft}
-                        variant={
-                          isRoot && isRound4
-                            ? matchTreePosition.centerSide
-                            : (side ?? PLAYER_POSITIONS.LEFT)
-                        }
-                        highlight={isRound2}
-                        shouldRender={
-                          isRoot && isRound4
-                            ? matchTreePosition.centerSide === PLAYER_POSITIONS.LEFT
-                              ? matchTreePosition.shouldRenderLeftLine
-                              : matchTreePosition.shouldRenderRightLine
-                            : true
-                        }
-                        round={currentRound}
-                      />
-                    )}
-                    <div className={styles.matchBox}>
-                      <PlayerCard
-                        player={match.player1}
-                        isWinner={match.player1?.id === match.winnerId}
-                        isLoser={isFinalRound && player1Eliminated}
-                        isReady={isPlayerReady(match.player1?.id, readyIds)}
-                        showStats={showStats}
-                        round={currentRound}
-                        side={
-                          isRoot && isRound4
-                            ? matchTreePosition.centerSide
-                            : (side ?? PLAYER_POSITIONS.LEFT)
-                        }
-                      />
-                      <span className={styles.vs}>vs</span>
-                      <PlayerCard
-                        player={match.player2}
-                        isWinner={match.player2?.id === match.winnerId}
-                        isLoser={isFinalRound && player2Eliminated}
-                        isReady={isPlayerReady(match.player2?.id, readyIds)}
-                        showStats={showStats}
-                        round={currentRound}
-                        side={
-                          isRoot && isRound4
-                            ? matchTreePosition.centerSide
-                            : (side ?? PLAYER_POSITIONS.LEFT)
-                        }
-                      />
-                    </div>
-                  </>
-                );
-              })()
-            )}
-          </div>
+          <div className={styles.centerColumn}>{renderCenterContent()}</div>
           {rightSemiFinal && (
             <MatchNode
               match={rightSemiFinal}
